Initial steps refactoring Player-App support
Adding static functions to the App class in order for the constructor to
implement an interface is silly and brittle. While it would be better to
avoid static properties entirely (perhaps we export an object that
contains the constructor etc.) this is a good step in the right
direction.
