+++ src/apps/prime/index.ts
@@ -1,13 +1,14 @@
 import _debug from "debug";
 const debug = _debug("babbling:PrimeApp");
 
-import { ChakramApi, ContentType, IBaseObj, ISeason } from "chakram-ts";
+import { ChakramApi } from "chakram-ts";
 
-import { IPlayableOptions, IQueryResult } from "../../app";
 import { ICastSession, IDevice } from "../../cast";
 import { BaseApp, MEDIA_NS } from "../base";
 import { awaitMessageOfType } from "../util";
+
 import { PrimeApi } from "./api";
+import { PrimePlayerChannel } from "./channel";
 import { IPrimeOpts } from "./config";
 
 export { IPrimeOpts } from "./config";
@@ -20,40 +21,9 @@ const DEFAULT_MARKETPLACE_ID = "ATVPDKIKX0DER";
 
 export class PrimeApp extends BaseApp {
 
-    public static ownsUrl(url: string) {
-        // TODO other domains
-        return url.includes("amazon.com");
-    }
-
-    public static async createPlayable(
-        url: string,
-        options: IPrimeOpts,
-    ) {
-        const titleId = pickTitleIdFromUrl(url);
-        if (!titleId) {
-            throw new Error(`Unsure how to play ${url}`);
-        }
-
-        const api = new ChakramApi(options.cookies);
-        const info = await api.getTitleInfo(titleId);
-        debug("titleInfo = ", info);
-
-        return playableFromObj(info);
-    }
-
-    public static async *queryByTitle(
-        title: string,
-        opts: IPrimeOpts,
-    ): AsyncIterable<IQueryResult> {
-        const api = new ChakramApi(opts.cookies);
-        for (const result of await api.search(title)) {
-            yield {
-                appName: "PrimeApp",
-                playable: playableFromObj(result),
-                title: cleanTitle(result.title),
-                url: "https://www.amazon.com/video/detail/" + result.id,
-            };
-        }
+    // declare Player support
+    public static createPlayerChannel() {
+        return new PrimePlayerChannel();
     }
 
     private readonly api: PrimeApi;
@@ -197,45 +167,4 @@ async function checkedRequest(session: ICastSession, message: any) {
     }
     debug(" -> ", resp);
     return resp;
-}
-
-function playableFromObj(info: IBaseObj) {
-    if (info.type === ContentType.SERIES) {
-        debug("playable for series", info.id);
-        return async (app: PrimeApp) => app.resumeSeries(info.id);
-    } else if (info.type === ContentType.SEASON) {
-        // probably they want to resume the series
-        const season = info as ISeason;
-        if (season.series) {
-            const seriesId = season.series.id;
-            debug("playable for series", seriesId, "given season", seriesId);
-            return async (app: PrimeApp) => app.resumeSeries(seriesId);
-        }
-    }
-
-    debug("playable for title", info.id);
-    return async (app: PrimeApp, opts: IPlayableOptions) => {
-        if (opts.resume === false) {
-            await app.play(info.id, { startTime: 0 });
-        } else {
-            await app.play(info.id, {});
-        }
-    };
-}
-
-function pickTitleIdFromUrl(url: string) {
-    const m1 = url.match(/video\/detail\/([^\/]+)/);
-    if (m1) {
-        return m1[1];
-    }
-
-    const m2 = url.match(/gp\/product\/([^\/\?]+)/);
-    if (m2) {
-        return m2[1];
-    }
-}
-
-function cleanTitle(original: string) {
-    // including this suffix confuses title-matching
-    return original.replace("(4K UHD)", "").trim();
 }
